+++ docs/reference/changelog.rst
@@ -21,6 +21,7 @@ Current version
 v5.5.0 - 2022-xx-xx
 -------------------
 
+- ``Changed`` Python patch bump + optimizations - by @goegol  [`#1665 <https://github.com/dsmrreader/dsmr-reader/issues/1665>`_]
 
 
 v5.4.0 - 2022-07-07